+++ lk_qtquick_scaffold/qml_helper/layout_helper.js
@@ -1,7 +1,55 @@
+function easyAlign(item, alignment) {
+    /*  Text.Align*
+     *       Text.AlignLeft: 1
+     *       Text.AlignHCenter: 4
+     *       Text.AlignRight: 2
+     *       Text.AlignTop: 32
+     *       Text.AlignBottom: 64
+     *       Text.AlignVCenter: 128
+     *  PS: TextInput.Align* are same with above.
+     *  See:
+     *      LCEdit
+     *      LCText
+     */
+    switch (alignment) {
+        case 'center':
+            item.horizontalAlignment = 4
+            item.verticalAlignment = 128
+            break
+        case 'hcenter':
+            item.horizontalAlignment = 4
+            break
+        case 'htop':
+            item.horizontalAlignment = 4
+            item.verticalAlignment = 32
+            break
+        case 'hbottom':
+            item.horizontalAlignment = 4
+            item.verticalAlignment = 64
+            break
+        case 'vcenter':
+            item.verticalAlignment = 128
+            break
+        case 'vleft':
+        // fall down
+        case 'lcenter':
+            item.horizontalAlignment = 1
+            item.verticalAlignment = 128
+            break
+        case 'vright':
+        // fall down
+        case 'rcenter':
+            item.horizontalAlignment = 2
+            item.verticalAlignment = 128
+            break
+    }
+}
+
+
 function autoWidth(parent) {
     let i, child
     let remainingWidth = parent.width -
-                         parent.p_spacing * (parent.children.length - 1)
+        parent.p_spacing * (parent.children.length - 1)
     let remainingChildren = Array()
 
     for (i in parent.children) {
@@ -15,7 +63,7 @@ function autoWidth(parent) {
         }
         remainingWidth -= child.width
     }
-    
+
     const eachRemainingChildWidth = remainingWidth / (remainingChildren.length)
     for (i in remainingChildren) {
         child = remainingChildren[i]
@@ -27,7 +75,7 @@ function autoWidth(parent) {
 function autoHeight(parent) {
     let i, child
     let remainingHeight = parent.height -
-                         parent.p_spacing * (parent.children.length - 1)
+        parent.p_spacing * (parent.children.length - 1)
     let remainingChildren = Array()
 
     for (i in parent.children) {