+++ lk_qtquick_scaffold/theme/LightClean/LCColumn.qml
@@ -7,11 +7,14 @@ Column {
     padding: LCDimension.Padding
     spacing: LCDimension.VSpacingM
 
-    property bool  p_alignCenter: false  // whether children aligned to horizontal center
+    property bool  p_alignCenter: false  // whether children aligned to
+    //      horizontal center
     property bool  p_autoHeight: true
-    // property bool  p_fillRest: false  // fill the last item to the remaining space
-    //      https://stackoverflow.com/questions/27319985/how-to-make-last-item-in-qml-container-fill-remaining-space
-    property bool  p_fillWidth: true
+    // property bool  p_fillRest: false  // fill the last item to the remaining
+    //      space
+    //      https://stackoverflow.com/questions/27319985/how-to-make-last-item
+    //      -in-qml-container-fill-remaining-space
+    property bool  p_fillWidth: false
     property alias p_padding: root.padding
     property alias p_spacing: root.spacing