+++ lk_qtquick_scaffold/theme/LightClean/LCEdit.qml
@@ -1,19 +1,28 @@
 import QtQuick 2.15
 import QtQuick.Controls 2.15
+
 import "./LCBackground"
 import "./LCStyle/dimension.js" as LCDimension
 import "./LCStyle/motion.js" as LCMotion
 import "./LCStyle/palette.js" as LCPalette
 import "./LCStyle/typography.js" as LCTypo
 
+import "../../qml_helper/layout_helper.js" as LCLayout
+
 TextField {
     id: root
-    leftPadding: LCDimension.HSpacingM; rightPadding: LCDimension.HSpacingM
+    implicitWidth: LCDimension.BarWidth
+    implicitHeight: LCDimension.BarHeight
+
+    leftPadding: LCDimension.HSpacingM
+    rightPadding: LCDimension.HSpacingM
     font.pixelSize: LCTypo.FontSizeM
     placeholderTextColor: LCPalette.TextHint
-    width: LCDimension.BarWidth; height: LCDimension.BarHeight
+    selectByMouse: true
+    selectedTextColor: LCPalette.TextSelected
+    selectionColor: LCPalette.TextSelection
 
-    /*  Tip: How to use validator in qml?
+    /*  Tip: How to use `validator` property?
             TextField {
                 validator: RegularExpressionValidator {
                     regularExpression: p_digitOnly ? /[0-9]+/ : /.* /
@@ -22,12 +31,8 @@ TextField {
         Note: This  wont be introduced into LCEdit, but you can do it
         yourself by following the tips above.
      */
-    
-    selectByMouse: true
-    selectedTextColor: LCPalette.TextSelected
-    selectionColor: LCPalette.TextSelection
 
-    property string p_alignment: 'center'
+    property string p_alignment: 'center'  // see `LCLayout.easyAlign`
     property alias  p_hint: root.placeholderText
     property alias  p_text: root.text
 
@@ -39,38 +44,6 @@ TextField {
 
     Component.onCompleted: {
         this.pressed.connect(this.clicked)
-    
-        switch (p_alignment) {  // modified based on `LCText`
-            case 'center':
-                this.horizontalAlignment = TextInput.AlignHCenter
-                this.verticalAlignment = TextInput.AlignVCenter
-                break
-            case 'hcenter':
-                this.horizontalAlignment = TextInput.AlignHCenter
-                break
-            case 'htop':
-                this.horizontalAlignment = TextInput.AlignHCenter
-                this.verticalAlignment = TextInput.AlignTop
-                break
-            case 'hbottom':
-                this.horizontalAlignment = TextInput.AlignHCenter
-                this.verticalAlignment = TextInput.AlignBottom
-                break
-            case 'vcenter':
-                this.verticalAlignment = TextInput.AlignVCenter
-                break
-            case 'vleft':
-                // fall down
-            case 'lcenter':
-                this.horizontalAlignment = TextInput.AlignLeft
-                this.verticalAlignment = TextInput.AlignVCenter
-                break
-            case 'vright':
-                // fall down
-            case 'rcenter':
-                this.horizontalAlignment = TextInput.AlignRight
-                this.verticalAlignment = TextInput.AlignVCenter
-                break
-        }
+        LCLayout.easyAlign(root, p_alignment)
     }
 }