+++ lk_qtquick_scaffold/theme/LightClean/LCListView.qml
@@ -4,26 +4,28 @@ import "./LCStyle/dimension.js" as LCDimension
 
 ListView {
     id: root
-    boundsBehavior: Flickable.DragOverBounds  // stop bouncing from scrolling to the start/end.
+    implicitWidth: LCDimension.FlowWidth
+    implicitHeight: LCDimension.FlowHeight
+
+    // stop bouncing from scrolling to the start/end.
+    boundsBehavior: Flickable.DragOverBounds
     clip: true
-    maximumFlickVelocity: 2300  // control mouse wheel velocity a little faster than default
+    // control mouse wheel velocity a little faster than default
+    maximumFlickVelocity: 2300
     model: p_model
     spacing: LCDimension.VSpacingXS
 
     property alias p_currentIndex: root.currentIndex
     property alias p_delegate: root.delegate
     property var   p_model: Array()  // [(str|dict), ...]
+    property alias p_scrollWidth: _scroll.width
     property alias p_spacing: root.spacing
     property int   r_count: p_model.length
     property alias r_currentItem: root.currentItem
 
     // show scroll bar
     // https://stackoverflow.com/questions/45650226/qml-attach-scrollbar-to-listview/45651291
-    ScrollBar.vertical: ScrollBar {}
-
-    Component.onCompleted: {
-        if (root.height == 0) {
-            root.height = childrenRect.height
-        }
+    ScrollBar.vertical: ScrollBar {
+        id: _scroll
     }
 }