+++ lk_qtquick_scaffold/theme/LightClean/LCRectangle.qml
@@ -4,9 +4,12 @@ import "./LCStyle/palette.js" as LCPalette
 
 Rectangle {
     id: root
-    border.width: 0; border.color: LCPalette.BorderNormal
+    implicitWidth: LCGeometry.BarWidth
+    implicitHeight: LCGeometry.BarHeight
+
+    border.width: 0
+    border.color: LCPalette.BorderNormal
     color: LCPalette.BgWhite
-    implicitWidth: LCGeometry.BarWidth; implicitHeight: LCGeometry.BarHeight
     radius: LCGeometry.RadiusS
 
     property alias p_border: root.border